Refactoring to become oled_display_node
