+++ launch/display.launch
@@ -1,3 +1,3 @@
 <launch>
-  <node pkg="display_node" type="display_node" name="display_node" output="screen" respawn="false" />
+  <node pkg="oled_display_node" type="oled_display_node" name="oled_display_node" output="screen" respawn="false" />
 </launch>
